requirements.txt
# Minimal requirements for production MD pipeline (host-side utils only)
pyyaml==6.0.1
tqdm==4.66.1
numpy==1.24.3
biopython==1.81
MDAnalysis==2.6.1
vermouth